refactor: 패키지 분리 및 컨벤션 통일
