+++ src/main/java/roomescape/controller/AdminController.java
@@ -1,17 +1,19 @@
-package roomescape;
+package roomescape.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
+@RequestMapping("/admin")
 public class AdminController {
-
-    @GetMapping("/admin")
+    
+    @GetMapping
     public String home() {
         return "admin/index";
     }
 
-    @GetMapping("/admin/reservation")
+    @GetMapping("/reservation")
     public String reservations() {
         return "admin/reservation-legacy";
     }
