+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,7 +1,5 @@
-package roomescape;
+package roomescape.controller;
 
-import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -13,17 +11,21 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import roomescape.domain.Reservation;
+import roomescape.dto.ReservationRequest;
+import roomescape.dto.ReservationResponse;
 
 @Controller
+@RequestMapping("/reservations")
 public class ReservationController {
 
-    private List<Reservation> reservations = new ArrayList<>();
+    private final List<Reservation> reservations = new ArrayList<>();
+    private final AtomicLong index = new AtomicLong(0);
 
-    private AtomicLong index = new AtomicLong(0);
-
-    @GetMapping("/reservations")
+    @GetMapping
     public ResponseEntity<List<ReservationResponse>> getReservations() {
-        List<ReservationResponse> reservationResponses = reservations.stream()
+        final List<ReservationResponse> reservationResponses = reservations.stream()
                 .map(reservation -> new ReservationResponse(
                         reservation.getId(),
                         reservation.getName(),
@@ -35,11 +37,11 @@ public class ReservationController {
         return ResponseEntity.ok(reservationResponses);
     }
 
-    @PostMapping("/reservations")
+    @PostMapping
     public ResponseEntity<ReservationResponse> createReservation(
-            @RequestBody ReservationRequest reservationRequest
+            @RequestBody final ReservationRequest reservationRequest
     ) {
-        Reservation reservation = new Reservation(
+        final Reservation reservation = new Reservation(
                 index.incrementAndGet(),
                 reservationRequest.name(),
                 reservationRequest.date(),
@@ -57,13 +59,12 @@ public class ReservationController {
         );
     }
 
-    @DeleteMapping("/reservations/{id}")
-    public ResponseEntity<Void> deleteReservation(@PathVariable("id") Long id) {
-        System.out.println();
-        Reservation reservation = reservations.stream()
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteReservation(@PathVariable("id") final Long id) {
+        final Reservation reservation = reservations.stream()
                 .filter(value -> Objects.equals(value.getId(), id))
                 .findFirst()
-                .orElseThrow(() ->  new IllegalArgumentException("아이디 없음"));
+                .orElseThrow(() -> new IllegalArgumentException("아이디 없음"));
         reservations.remove(reservation);
 
         return ResponseEntity.ok().build();
